requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
flask-cors==4.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
mysqlclient==2.2.0
pymysql==1.0.3
SpeechRecognition==3.10.0
gunicorn==21.2.0
python-dotenv==1.0.0
reportlab==4.0.4
pandas==2.0.3
openpyxl==3.1.2
PyPDF2==3.0.1
pillow==10.0.0
bcrypt==4.0.1
cryptography==41.0.4
blinker==1.6.2
itsdangerous==2.1.2
click==8.1.3
MarkupSafe==2.1.3
